meta-nexell-distro: remove alsa-state and alsa-restore
This patch removes alsa-state and alsa-restore feature.

Previously, the alsa-state and alsa-restore service files were removed from
the post-processing phase, but this is not the right approach. So we changed
the method to exclude alsa-state and alsa-restore which are included in actual alsa-utils.

There is a bug that turns off the audio mixer intermittently by the alsa-state
and alsa-restore services that are running as systemd services,
and it is not absolutely necessary in our system. They will not be used in the future.

Signed-off-by: Lee Geunjeong <kjlee@nexell.co.kr>
Change-Id: If22b1b68c972171a981dac6528491548dcec565f